First commit. Started optimising the code so that the structure makes more sense. There will be more nesting of elements and renaming in the next commits
