+++ src/webparts/ccs/CcsWebPart.ts
@@ -6,14 +6,13 @@ import {
   PropertyPaneTextField
 } from '@microsoft/sp-property-pane';
 import { BaseClientSideWebPart } from '@microsoft/sp-webpart-base';
-
 import * as strings from 'CcsWebPartStrings';
 import Ccs from './components/Ccs';
-import { ICcsProps } from './components/ICcsProps';
 
-import { mockArray } from './data/QuestionData';
 import { Environment, EnvironmentType } from '@microsoft/sp-core-library';
 import { SPHttpClient, SPHttpClientResponse } from '@microsoft/sp-http';
+import { ICcsProps } from './components/ICcsProps';
+import { mockArray } from './mockData/FormData';
 
 export interface ICcsWebPartProps {
   description: string;
