+++ src/webparts/ccs/components/FormElements/IForm.ts
@@ -1,0 +1,7 @@
+export interface IFormProps {
+  arrayToUse: any[];
+  }
+  
+export interface IFormState {
+  arrayToUse: any[];
+}
